backend_fastapi/requirement.txt
pandas
Pillow
matplotlib
numpy
fastapi
uvicorn
python-multipart
opencv-python
transformers
tensorflow